requirements.txt
pandas 
numpy 
tensorflow 
scikit-learn
matplotlib 
opencv-python
fastapi
pillow
uvicorn
python-multipart
